add paypal payment entity: amount entity and deatls

+++ lib/constants.dart
@@ -5,7 +5,7 @@ const double kSizedBoxHeight16 = 16.0;
 const double kSizedBoxHeight33 = 33.0;
 const String kIsOnBoardingViewSeen = 'IsOnBoardingViewSeen';
 const String kSaveDataSharedPref = "saveData";
-const int kPayWithCashCost = 40;
+const double kPayWithCashCost = 40;
 const EdgeInsetsGeometry kPrimaryScreenPadding = EdgeInsets.only(
   top: 16,
   right: 16,
